+++ src/onchain-sample-contracts.js
@@ -5,6 +5,8 @@ export const NETWORKS_BY_CHAIN_ID = {
   56: 'bsc',
   43114: 'avalanche',
   42161: 'arbitrum',
+  11155111: 'sepolia',
+  8453: 'base',
 };
 
 export const ERC20_SAMPLE_CONTRACTS = {
@@ -13,6 +15,8 @@ export const ERC20_SAMPLE_CONTRACTS = {
   bsc: '0x8965349fb649A33a30cbFDa057D8eC2C48AbE2A2',
   avalanche: '0xB97EF9Ef8734C71904D8002F8b6Bc66Dd9c48a6E',
   arbitrum: '0xaf88d065e77c8cC2239327C5EDb3A432268e5831',
+  sepolia: '0x27A56df30bC838BCA36141E517e7b5376dea68eE',
+  base: '0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913',
 };
 
 export const ERC721_SAMPLE_CONTRACTS = {
@@ -21,4 +25,6 @@ export const ERC721_SAMPLE_CONTRACTS = {
   bsc: '0xebfbfd7c41b123500fb16b71c43b400c12b08be0',
   avalanche: '0x880f7e04D722e305126F7E1efd3434A7d5b1465c',
   arbitrum: '0x8659a4876369b94515a86048fe7f99daba6b9a7d',
+  sepolia: '0xbba60aa8144579e07c6db64121b0f608ab6f0c89',
+  base: '0x90997fc967e75b7e69f899133aab31d197beb802',
 };